cập nhật xử lý nghiệp vụ

+++ WebKhoaLuan/src/main/java/com/demo/WebKhoaLuan/controller/HoidongController.java
@@ -6,8 +6,10 @@ package com.demo.WebKhoaLuan.controller;
 
 import com.demo.WebKhoaLuan.model.Chitiethoidong;
 import com.demo.WebKhoaLuan.model.Hoidong;
+import com.demo.WebKhoaLuan.model.Khoaluan;
 import com.demo.WebKhoaLuan.repository.ChitiethoidongRepository;
 import com.demo.WebKhoaLuan.repository.HoidongRepository;
+import com.demo.WebKhoaLuan.repository.KhoaluanRepository;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -29,32 +31,40 @@ public class HoidongController {
     private HoidongRepository hoiDongRepository;
     @Autowired
     private ChitiethoidongRepository chiTietHoiDongRepository;
+    @Autowired
+    private KhoaluanRepository KhoaluanRepository;
     
+    //GIÁO VỤ LẤY DANH SÁCH HỘI ĐỒNG
     @GetMapping("/giaovu/dsHoiDong")
     public List<Hoidong> dsHoiDong(){
         return hoiDongRepository.findAll();
     }
     
+    //GIÁO VỤ XEM MỘT HỘI ĐỒNG
     @GetMapping("/giaovu/dsHoiDong/{maHd}")
     public Hoidong layHoiDong(@PathVariable(value = "maHd") int maHd){
         return hoiDongRepository.layHD(maHd);
     }
     
+    //GIÁO VỤ LẤY DANH SÁCH HỘI ĐỒNG HOẠT ĐỘNG 
     @GetMapping("/giaovu/dsHDHoatDong")
     public List<Hoidong> dsHDHoatDong(){
         return hoiDongRepository.dsHDHoatDong(Short.parseShort("1"));
     }
     
+    //GIÁO VỤ LẤY DANH SÁCH HĐ ĐỒNG HẾT HOẠT ĐỘNG
     @GetMapping("/giaovu/dsHDKhongHoatDong")
     public List<Hoidong> dsHDKHoatDong(){
         return hoiDongRepository.dsHDHoatDong(Short.parseShort("0"));
     }
     
+    //GIẢNG VIÊN LẤY DANH SÁCH HỘI ĐỒNG GIANG VIÊN THAM GIA
     @GetMapping("/giangvien/dsHoiDongGV/{maGv}")
     public List<Chitiethoidong> dsHDGV(@PathVariable(value = "maGv") String maGv){
         return hoiDongRepository.layHDGV(maGv);
     }
     
+    //GIAO VỤ CẬP NHẬT TÌNH TRẠNG CỦA HỘI ĐỒNG
     @PostMapping("/giaovu/tinhTrangHD/{maHd}")
     public String tinhTrangHD(@PathVariable(value = "maHd") int maHd){
         Hoidong hd = hoiDongRepository.layHD(maHd);
@@ -70,6 +80,7 @@ public class HoidongController {
         return "Cập nhật thành công";
     }
     
+    //GIÁO VỤ THÊM HỘI ĐỒNG MỚI
     @PostMapping("/giaovu/themHD")
     public String themHD(@RequestBody Hoidong hoidong){
         Hoidong hd = new Hoidong();
@@ -81,25 +92,45 @@ public class HoidongController {
         return "Thêm hội đồng thành công";
     }
     
+    //GIÁO VỤ PHÂN CÔNG GIẢNG VIÊN CHO HỘI ĐỒNG 
     @PostMapping("/giaovu/phanCong/{maHd}")
     public String phanCong(@PathVariable(value = "maHd") int maHd, @RequestBody Chitiethoidong chitiethoidong){
-        Hoidong hd = hoiDongRepository.layHD(maHd);
-        chitiethoidong.setHoidong(hd);
-        try {
-            chiTietHoiDongRepository.save(chitiethoidong);
-        } catch (Exception e) {
-            return "Phân công giảng viên không thành công";
+        if (kiemSLThanhVien(maHd) < 5){
+            Hoidong hd = hoiDongRepository.layHD(maHd);
+            chitiethoidong.setHoidong(hd);
+            try {
+                chiTietHoiDongRepository.save(chitiethoidong);
+            } catch (Exception e) {
+                return "Phân công giảng viên không thành công";
+            }
+            return "Phân công giảng viên thành công";
         }
-        return "Phân công giảng viên thành công";
+        return "Hội đồng đã đủ thành viên, không thể gán thêm";
     }
     
+    //GIÁO VỤ XÓA HỘI ĐỒNG
     @DeleteMapping("/giaovu/xoaHD/{maHd}")
     public String xoaHD(@PathVariable(value = "maHd") int maHd){
-        try {
-            hoiDongRepository.deleteById(maHd);
-        } catch (Exception e) {
-            return "Xóa hội đồng không thành công";
+        if (kiemSLKhoaLuan(maHd) == 0) {
+            try {
+                hoiDongRepository.deleteById(maHd);
+            } catch (Exception e) {
+                return "Xóa hội đồng không thành công";
+            }
+            return "Xóa hội đồng thành công";
         }
-        return "Xóa hội đồng thành công";
+        return "Không được phép xóa hội đồng đã có khóa luận";
+    }
+
+    //KIỂM TRA SỐ LƯỢNG THÀNH VIÊN MỘT HỘI ĐỒNG
+    public int kiemSLThanhVien(int maHd){
+        List<Chitiethoidong> ds = hoiDongRepository.kiemSLTV(maHd);
+        return ds.size();
+    }
+    
+    //KIỂM TRA SỐ LƯỢNG KHÓA LUẬN MỘT HỘI ĐỒNG
+    public int kiemSLKhoaLuan(int maHd){
+        List<Khoaluan> dsKL = KhoaluanRepository.layDsKLHD(maHd);
+        return dsKL.size();
     }
 }
